+++ src/Symfony/Component/HttpKernel/EventListener/ErrorListener.php
@@ -164,9 +164,9 @@ class ErrorListener implements EventSubscriberInterface
      * 
      * @param ?string $logChannel
      */
-    protected function logException(\Throwable $exception, string $message, ?string $logLevel = null, /* , ?string $logChannel = null */): void
+    protected function logException(\Throwable $exception, string $message, ?string $logLevel = null, /* ?string $logChannel = null */): void
     {
-        $logChannel = 3 < \func_num_args() ? \func_get_arg(3) : $this->resolveLogChannel($exception);
+        $logChannel = (3 < \func_num_args() ? \func_get_arg(3) : null) ?? $this->resolveLogChannel($exception);
 
         $logLevel ??= $this->resolveLogLevel($exception);